backend/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
djangorestframework-simplejwt
Pillow